Add cayleys closure condition visualization & settings
Adds a visualization for cayleys closure condition as well as settings to
change the size of the ellipses & the displayed functions.

The python files contain a script to generate the implicit function for
cayleys closure condition.

This commit also fixes a bug in Firefox when left-clicking that prevented
the used from dragging the inner ellipse or the starting point